refactor the way a media item is turned into a list of playlist items
An item can now act as a Streamable which provides functionality for
returning an array of ItemLink objects. The PlaylistItem constructor
has been modified to take an ItemLink as parameter and deduce all
necessary information from it. This enables us to reuse all of the
playlist-related code, and it makes it easier to add functionality
for creating playlists for complete TV shows.