refactor the user settings so that every user can change his own settings
This introduces a new setting to choose the default landing page after login. It also moves the selection of language to the user settings page, so the change language modal has been removed. There is also no longer a concept of a temporary interface language change, so that part of the code has been simplfied somewhat.

Closes #266