Sort movies and TV shows by "premiered" instead of "year"
This also changes the rendering of the year to use the year from the premiere date when it's available, otherwise the "year" column, and if nothing is available, "Not available" is returned. This also fixes a bug where 1601 is shown as the year for items that don't apparently have a year (Kodi sometimes returns 1601 instead of 0).

Closes #317